test_background: change test media
The crater_lake.png was "free beer" imagery, but the usage terms were not clear
on distribution, so removing.

So, use one my personal snaps (bridge_of_the_gods.png) from Skamania Lodge,
Washington, taken with my smartphone on a gorgeous morning in March 2012. I
relinquish copyright of the image, and hereby release it to the public domain.

Signed-off-by: Joe Konno <joe.konno@intel.com>