EFL: Remove Popup resize test
Although a Popup is an Evas_Object, I'm not sure where E upstream wants to take
this control, in light of recent Popup-Resize test failures. In most popup
schemes with which I'm familiar (growl, libnotify), the particulars of popups
are often opaque, intended for "drop-in-and-use" with very little granular
control from the developer's perspective.

If someone wants to write their own popup sized just so, with chrome bells and
whistles, they are free to use the Evas API, as far as I'm concerned.

Signed-off-by: Joe Konno <joe.konno@intel.com>