efl: refactor getSurfaceGeometry and fix window resize on wayland_shm
Remove need for a callback in ElmTestHarness::getSurfaceGeometry().  The old
method only used the step queue to gain a yield when waiting for the request
to finish.  However, we can just yield directly.

On wayland_shm, the window titlebar animation runs right after the window is
shown and seems to affect resize operations.  Therefore, just add enough
yield's to allow the animation to finish before applying resize operations.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>