efl: quick workaround to fix issue with recent efl commit
<http://git.enlightenment.org/core/efl.git/commit/?id=a37443ad879815fb1c327a7a1b6669a84e8cf622>
removed the wl_display_dispatch() in the ecore_wl_init() function.
Thus we can longer rely on it to trigger the global registry
callbacks to populate the globals before any objects are created.
As a quick workaround we just bind to the wfits protocol on first
use and just assume the globals will be available by then.  We
should probably use a better solution however.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>