weston-wfits: track active keys and deactivate on client disconnect
Test clients might disconnect before deactivating any keys that it
has activated and affect other tests that run after it.  For
instance, sending a pointer button down and then exiting causes
the mouse to be stuck in a down state in the compositor.

This patch adds functionality to the weston-wfits test extension to
track active keys and deactivates any keys that are still active
when a client disconnnects from the compositor.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>