weston-wfits: split logic into separate files
Split the logic for the different wfits interfaces into
separate files.  Used classes to encapsulate logic but was forced
to use primarily static members/functions since Weston has a
bug that results in a detrimental side-effect of destructing
objects that are loaded when it fails to fork/exec client
processes (e.g. weston-keyboard, weston-screensaver, etc.).  That
is, when Weston fails to fork/exec it calls exit(-1) instead
of _exit(-1) and hence invokes atexit and destructors for the
cloned objects in its forked process.  Those destructors might
unload/uninitialize things that effect the main process (e.g.
sockets, ioctls, etc...).

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>