weston-wfits: Allow user to configure the input emulator
The available emulators are "uinput" and "notify".  To choose
which emulator to use, define the environment variable:

WESTON_WFITS_INPUT_EMULATOR

equal to the name of the desired emulator before loading the
weston-wfits.so module into Weston.

For example:

 $ export WESTON_WFITS_INPUT_EMULATOR="notify"
 $ weston --modules=weston-wfits.so

If WESTON_WFITS_INPUT_EMULATOR is not defined in the environment
and Weston is using the headless-backend then weston-wfits.so will
choose "notify" by default.  Otherwise, it chooses "uinput" by
default.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>