weston-wfits: use unique timestamp for notify*
Use weston_compositor_get_time() instead of a hard-coded value of
100 for the timestamp in each call to a Weston notify method.  This
ensures the probability that each notify event has a unique
timestamp.  Some clients may opt to ignore events when it has
already processed a similar event with the same timestamp.  Thus,
hardcoding the time value is not a good idea.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>