weston-wfits.so: support 1.4 weston sdk
Weston surface geometry is now tracked by weston views.  Update
the wfits query interface to support the new structure.  For now,
only support surfaces with a single view.

The "focus" flag is no longer a member of the weston compositor as
it is not needed anymore.  This was originally used to ensure notify
api's always worked when used on x11-backend if the compositor window
didn't have X focus.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>