efl/harness: don't use a static function variable for the client
When the program exits, there seems to be a race between when
wl_display_disconnect is called by ecore_wl_shutdown (i.e. when the
Application instance is destructed) and when the static client instance
is destructed.  The client instance destroys some bound interface
objects that require a valid display connection which may be gone
by the time the client instance is destroyed.  Hence, resulting
in a segfault.

We still need to create the client instance on first use since
ecore_wl_display_get() might not be available during the harness
instantiation.  Thus, just manage the client_ instance as a member
pointer in the ElmTestHarness instance and delete it when the
harness destructs... which will be before the Application instance
is destroyed.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>