Make efl tests work with g-main-loop enabled efl
This is a major refactor of the efl test framework
to get the efl tests to work with a g-main-loop
enabled efl.

The efl test framework now uses a multi-threaded
model (as opposed to the recursive iterate model) to
drive the tests.  That is, the efl main loop executes
in the main thread and the tests execute in a separate
thread.  All calls into the efl api's are synchronized
with the main loop by employing the efl threading
synchronization api's.

Fixes #6

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>