efl/ListUserMouseScrollTest: move mouse the length of surface
With some elm profiles, the list items are taller/wider.  The list
only scrolls the distance of the pointer relative to the list item.
With larger sized list items, the previous absolute hard-coded delta
only generated small list movements resulting in the test taking
longer to scroll the list to its edges; ultimately the test timeout
wins before the test completes and the test fails.

Now, move the pointer over the entire length/height of the window
to ensure larger list scrolls.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>