gtk: fix fullscreen test
We assumed that the surface is fullscreened right away we send
fullscreen request. This may not be true, since events go like this:

    client              server
-----------------------------------
 -> fullscreen
                     <- configure
 -> set_geometry
 -> < draw, commit >

We checked for client's size after set_geometry but before
drawing and commiting, so the size was not right yet. Also
fullscreening may have an animation, so the client can have
some intermediate sizes before it is resized over whole screen.
The same when unfullscreening back to initial size.

Fix it by waiting until the client has the fullscreen (initial) size.
(If it won't, the test will time-out)

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>