build: Support out-of-tree compilation
Some build systems, such as Yocto, use a separate directory
to output compilation results.

As we are creating some source and headers files with the
help of "wayland-scanner", those files will end up in this
dedicated build directory ; in order for the process to
succeed, it is thus needed that the Makefiles reference it.

Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>