build: Support building with Weston >= 1.8.91
In recent versions of Weston, the "container_of" macro  is
not exposed via a public header anymore. Just redefine it
locally if needed; it keeps the code readable and will
always work.

Since Weston 1.8.91 (well more precisely, somewhere in the
middle of 1.8.90 but we have no way to test that) member
elements of "weston_seat" (such as "weston_pointer",
"weston_touch"...) can not be accessed directly anymore,
but require a dedicated function. Since it is only
required in one place and wayland-fits is all about testing
various versions of Wayland/Weston, let us have a #ifdef
until such breakages begin to happen everywhere.

Change-Id: If3bf867133c86ac43c5e5d7031cb467d5109c7b9
Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>